Sam/Kris - #55 - fixing bad stop script
git-svn-id: svn://192.168.0.100/project/trunk/bloodtorrent@49 363a94d9-e023-4002-b4ac-95dfb99b3300

+++ scripts/stopDropwizard.sh
@@ -2,9 +2,11 @@
 
 PIDFILE=/var/run/bloottorrent.pid
 
-PID=`cat $PIDFILE`
-if [ ! -z "$PID" ] ; then
-    kill -9 $PID
+if [ -f $PIDFILE ] ; then
+	PID=`cat $PIDFILE`
+	if [ ! -z "$PID" ] ; then
+	    kill -9 $PID
+	fi
+	rm $PIDFILE
 fi
 
-rm $PIDFILE
